requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.66.2
WooCommerce==3.0.0
